+++ myorderapp/src/main/java/za/ac/cput/myorderapp/conf/Factory/ItemFactory.java
@@ -9,7 +9,7 @@ import java.util.Map;
  * Created by student on 2015/04/25.
  */
 public class ItemFactory {
-    public static Item createItem(int itemNo, Map<String, String> values, TotalPrice price){
+    public static Item createItem(Long itemNo, Map<String, String> values, TotalPrice price){
         Item item = new Item.Builder(values.get("name"))
                 .itemNo(itemNo)
                 .brand(values.get("brand"))
