+++ myorderapp/src/main/java/za/ac/cput/myorderapp/conf/Factory/SupplierFactory.java
@@ -8,7 +8,7 @@ import java.util.Map;
  * Created by student on 2015/04/25.
  */
 public class SupplierFactory {
-    public static Supplier createSupplier(int supplierNo, Map<String, String> values){
+    public static Supplier createSupplier(Long supplierNo, Map<String, String> values){
         Supplier supplier = new Supplier.Builder(values.get("name"))
                 .phone("phone")
                 .supplierNo(supplierNo)
